requirements.txt
Django>=3.0,<4.0
psycopg2>=2.7,<3.0
python-dotenv>=0.11,<1.0
Pillow>=7.0,<8.0
djangorestframework>=3.11,<4.0
djangorestframework-simplejwt>=4.4,<5.0
django-cors-headers>=3.2,<4